Add files via upload
new goto line number, cleaned up break-point counter to better enable go number of breakpoints to be accurate even if reporting is turned off. Provides a way to re-run a program and stop short of where you found a problem, only works if program is totally repeatable, but provides a poor-mans run backwards capability (well, that's what I use it for).  Instrumentation now has an enable/disable for cases where it doesn't work right, or you want to turn off single stepping for a segment of code, e.g. a long loop. The manual has been updated as well.