Add files via upload
some cleanup for instrumentation of multi-line statements, such as set or puts. Now also checks for procs and methods that don't use a return statement to return a value. An optional warning will be sent to stderr if this is found to be the case, as this could affect the return value by instrumenting.